first sample of sensor streaming with individual package (gyroscope)
using cordova-plugin-gyroscope from @jasonyang, I'm able to fetch data from the gyroscope in my device, with a watching method. but the thing is, it's no good if we have to install an individual plugin for every single sensor, mainly because they're not consistent about which devices they'll work on (for example, the cordova-plugin-altimeter from @jameslegue only works on iOS). next time, we'll try to centralize the sensor's calls with only one plugin  (using cordova-plugin-sensors from @fabiorogeriosj, and later on modifying it for more complete services).